+++ src/Sql.Cmdlets/Get-First.cs
@@ -23,6 +23,12 @@ public class GetFirstCommand: PSCmdlet {
 	public required string Command { get; set; }
 
 	/// <summary>
+	/// Value indicating how the command is interpreted.
+	/// </summary>
+	[Parameter]
+	public CommandType CommandType { get; set; } = CommandType.Text;
+
+	/// <summary>
 	/// The connection to the data source.
 	/// </summary>
 	[Parameter(Mandatory = true, Position = 0)]