+++ src/Sql/ListExtensions.cs
@@ -4,7 +4,7 @@ namespace Belin.Sql;
 /// Provides extension members for lists.
 /// </summary>
 public static class ListExtensions {
-	// TODO (.NET 10) extension(IDbConnection connection)
+	// TODO (.NET 10) extension(IList list)
 
 	/// <summary>
 	/// Converts the specified list into an ordered dictionary.
@@ -13,5 +13,5 @@ public static class ListExtensions {
 	/// <param name="prefix">The prefix to use for generating the dictionary key names.</param>
 	/// <returns>The ordered dictionary corresponding to the specified list.</returns>
 	public static OrderedDictionary<string, T?> ToOrderedDictionary<T>(this IList<T?> list, string prefix = "") =>
-		new(list.Index().Select(tuple => new KeyValuePair<string, T?>($"{prefix}{tuple.Index}", tuple.Item)));
+		new(list.Index().Select(entry => new KeyValuePair<string, T?>($"{prefix}{entry.Index}", entry.Item)));
 }