+++ ChangeLog.md
@@ -3,8 +3,8 @@
 ## Version [0.8.0](https://github.com/cedx/sql.net/compare/v0.7.0...v0.8.0)
 - Moved the data mapping to the `Belin.Sql` assembly.
 - Moved the cmdlets to the `Belin.Sql.Cmdlets` assembly.
-- Added the `-Type` parameter to the `New-Command` cmdlet.
 - Added the `Belin.Sql.Dapper` assembly providing type maps and type handlers for [Dapper](https://www.learndapper.com).
+- Added the `-CommandType` parameter to most cmdlets.
 - Added the `-DbType` and `-Direction` parameters to the `New-Parameter` cmdlet.
 
 ## Version [0.7.0](https://github.com/cedx/sql.net/compare/v0.6.0...v0.7.0)